+++ fastchart_arginfo.h
@@ -1,5 +1,5 @@
 /* This is a generated file, edit the .stub.php file instead.
- * Stub hash: 206fc1089f4ff900e2c1fd6e023078cdb0a84549 */
+ * Stub hash: 39d6deacaa8afda079f40606ddefb011545a0074 */
 
 ZEND_BEGIN_ARG_WITH_RETURN_TYPE_INFO_EX(arginfo_class_FastChart_Chart_version, 0, 0, IS_STRING, 0)
 ZEND_END_ARG_INFO()
@@ -23,6 +23,10 @@ ZEND_END_ARG_INFO()
 
 ZEND_BEGIN_ARG_WITH_RETURN_TYPE_INFO_EX(arginfo_class_FastChart_Chart_setFontSize, 0, 1, IS_STATIC, 0)
 	ZEND_ARG_TYPE_INFO(0, size, IS_DOUBLE, 0)
+ZEND_END_ARG_INFO()
+
+ZEND_BEGIN_ARG_WITH_RETURN_TYPE_INFO_EX(arginfo_class_FastChart_Chart_setCategoryLabels, 0, 1, IS_STATIC, 0)
+	ZEND_ARG_TYPE_INFO(0, labels, IS_ARRAY, 0)
 ZEND_END_ARG_INFO()
 
 ZEND_BEGIN_ARG_WITH_RETURN_OBJ_INFO_EX(arginfo_class_FastChart_Chart_draw, 0, 1, GdImage, 0)
@@ -79,6 +83,7 @@ ZEND_METHOD(FastChart_Chart, setTitle);
 ZEND_METHOD(FastChart_Chart, setTheme);
 ZEND_METHOD(FastChart_Chart, setFontPath);
 ZEND_METHOD(FastChart_Chart, setFontSize);
+ZEND_METHOD(FastChart_Chart, setCategoryLabels);
 ZEND_METHOD(FastChart_LineChart, setSeries);
 ZEND_METHOD(FastChart_LineChart, draw);
 ZEND_METHOD(FastChart_BarChart, setSeries);
@@ -101,6 +106,7 @@ static const zend_function_entry class_FastChart_Chart_methods[] = {
 	ZEND_ME(FastChart_Chart, setTheme, arginfo_class_FastChart_Chart_setTheme, ZEND_ACC_PUBLIC)
 	ZEND_ME(FastChart_Chart, setFontPath, arginfo_class_FastChart_Chart_setFontPath, ZEND_ACC_PUBLIC)
 	ZEND_ME(FastChart_Chart, setFontSize, arginfo_class_FastChart_Chart_setFontSize, ZEND_ACC_PUBLIC)
+	ZEND_ME(FastChart_Chart, setCategoryLabels, arginfo_class_FastChart_Chart_setCategoryLabels, ZEND_ACC_PUBLIC)
 	ZEND_RAW_FENTRY("draw", NULL, arginfo_class_FastChart_Chart_draw, ZEND_ACC_PUBLIC|ZEND_ACC_ABSTRACT, NULL, NULL)
 	ZEND_FE_END
 };